+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
+## [Unreleased]
+
+- [unreleased]: https://github.com/rogdham/bigxml/compare/v0.10.0...HEAD
+
+### :rocket: Added
+
+- Support `functools.partial` handlers
+
 ## [0.10.0] - 2023-04-22
 
 [0.10.0]: https://github.com/rogdham/bigxml/compare/v0.9.0...v0.10.0